refactor: extract log file download handling out of PluginLogger
