+++ plugin/webpay-rest.php
@@ -6,6 +6,7 @@ use Transbank\WooCommerce\WebpayRest\Controllers\TransactionStatusController;
 use Transbank\WooCommerce\WebpayRest\Helpers\DatabaseTableInstaller;
 use Transbank\WooCommerce\WebpayRest\Helpers\HposHelper;
 use Transbank\WooCommerce\WebpayRest\Helpers\TbkFactory;
+use Transbank\Plugin\Helpers\LogFileDownloadHandler;
 use Transbank\WooCommerce\WebpayRest\PaymentGateways\WC_Gateway_Transbank_Oneclick_Mall_REST;
 use Transbank\WooCommerce\WebpayRest\PaymentGateways\WC_Gateway_Transbank_Webpay_Plus_REST;
 use Transbank\WooCommerce\WebpayRest\Blocks\WCGatewayTransbankWebpayBlocks;
@@ -13,7 +14,6 @@ use Transbank\WooCommerce\WebpayRest\Blocks\WCGatewayTransbankOneclickBlocks;
 use Transbank\WooCommerce\WebpayRest\Setup\ConfigMigrator;
 use Transbank\WooCommerce\WebpayRest\Utils\ConnectionCheck;
 use Transbank\WooCommerce\WebpayRest\Utils\TableCheck;
-use Transbank\Plugin\Helpers\PluginLogger;
 use Transbank\WooCommerce\WebpayRest\Utils\Template;
 use Automattic\WooCommerce\Internal\DataStores\Orders\CustomOrdersTableController;
 use Transbank\WooCommerce\WebpayRest\Admin\Notices\AdminNoticeManager;
@@ -59,8 +59,8 @@ add_action('add_meta_boxes', function () use ($hposExists) {
 
 add_action('init', function () {
     add_action('wp_ajax_check_connection', ConnectionCheck::class . '::check');
-    add_action('wp_ajax_check_can_download_file', PluginLogger::class . '::checkCanDownloadLogFile');
-    add_action('wp_ajax_download_log_file', PluginLogger::class . '::downloadLogFile');
+    add_action('wp_ajax_check_can_download_file', LogFileDownloadHandler::class . '::checkCanDownloadLogFile');
+    add_action('wp_ajax_download_log_file', LogFileDownloadHandler::class . '::downloadLogFile');
     add_action('wp_ajax_get_transaction_status', [new TransactionStatusController(), 'getStatus']);
 });
